+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Updates
 
 ### Fixed
+- Clear optional tail in getvalue. See #1055 #1065
 
 ## [3.2.2] - 2015-09-04